Rename file to base and change class name of ips, snips
