+++ estimators/slates/base.py
